requirements.txt
torch==1.3.1
torchvision==0.4.2
numpy==1.17.2
Pillow==6.2.0
nibabel==3.2.1
opencv-python==4.1.2.30
pandas==0.25.1
ttach==0.0.3